shop/requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.5.0
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==21.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.1.0
blinker==1.4
Brlapi==0.8.3
celery==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
command-not-found==0.3
cryptography==41.0.4
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.8.0rc2
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.6
django-autoslug==1.9.9
django-debug-toolbar==4.2.0
django-environ==0.10.0
django-js-asset==2.1.0
django-mail-templated==2.6.5
django-mptt==0.14.0
django-phonenumber-field==7.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.20
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.7
filelock==3.6.0
frozenlist==1.4.0
greenlet==2.0.2
gunicorn==20.1.0
gyp==0.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
inflection==0.5.1
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.5.0
kombu==5.3.1
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
magic-filter==1.0.11
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.2
olefile==0.46
packaging==23.2
paramiko==3.3.1
pexpect==4.8.0
phonenumberslite==8.13.17
Pillow==10.0.1
platformdirs==2.5.1
progressbar2==4.2.0
prompt-toolkit==3.0.39
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.3.0
pydantic_core==2.6.3
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-utils==3.8.1
python3-openid==3.2.0
pytils==0.4.1
pytz==2023.3.post1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
soupsieve==2.5
SQLAlchemy==2.0.20
sqlparse==0.4.4
systemd-python==234
termcolor==2.3.0
texttable==1.6.4
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==2.0.6
vine==5.0.0
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.6
websocket-client==0.59.0
WoeUSB-ng==0.2.12
wxPython==4.0.7
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
